+++ src/base/tests/complex-spec.js
@@ -2,17 +2,17 @@
  *  complext tc for base and attribute
  *  @author yiminghe@gmail.com
  */
-KISSY.use("base", function(S, Base) {
-    describe("complex base/attribute", function() {
+KISSY.use("base", function (S, Base) {
+    describe("complex base/attribute", function () {
 
-        it("can merge property value object from parent class", function() {
+        it("can merge property value object from parent class", function () {
             function a() {
                 a.superclass.constructor.apply(this, S.makeArray(arguments));
             }
 
             a.ATTRS = {
                 x:{
-                    getter:function() {
+                    getter:function () {
                         return 1;
                     }
                 }
@@ -39,7 +39,7 @@ KISSY.use("base", function(S, Base) {
         });
 
 
-        it("support validator", function() {
+        it("support validator", function () {
 
             function a() {
                 a.superclass.constructor.apply(this, S.makeArray(arguments));
@@ -47,7 +47,7 @@ KISSY.use("base", function(S, Base) {
 
             a.ATTRS = {
                 tt:{
-                    validator:function(v) {
+                    validator:function (v) {
                         return v > 1;
                     }
                 }
@@ -67,7 +67,77 @@ KISSY.use("base", function(S, Base) {
 
         });
 
-        it("support sub attribute name", function() {
+
+        it("support validators", function () {
+
+            function Aa() {
+                Aa.superclass.constructor.apply(this, S.makeArray(arguments));
+            }
+
+            Aa.ATTRS = {
+                tt:{
+                    validator:function (v, name,all) {
+                        if (all && (v > all["t"])) {
+                            return "tt>t!";
+                        }
+                    }
+                },
+                t:{
+                    validator:function (v) {
+                        if (v < 0) {
+                            return "t<0!";
+                        }
+                    }
+                }
+
+            };
+
+            S.extend(Aa, Base);
+
+            var t = new Aa(),
+                e1;
+
+            expect(t.set("t", -1, {
+                error:function (v) {
+                    e1 = v;
+                }
+            })).toBe(false);
+
+            expect(e1).toBe("t<0!");
+            expect(t.get("t")).not.toBe(-1);
+
+            var e2;
+
+            expect(t.set({
+                tt:2,
+                t:-1
+            }, {
+                error:function (v) {
+                    e2 = v;
+                }
+            })).toBe(false);
+
+            expect(e2.sort()).toEqual(["t<0!", "tt>t!"].sort());
+            expect(t.get("t")).not.toBe(-1);
+            expect(t.get("tt")).not.toBe(2);
+
+            var e3;
+            expect(t.set({
+                tt:3,
+                t:4
+            }, {
+                error:function (v) {
+                    e3 = v;
+                }
+            })).not.toBe(false);
+
+            expect(e3).toBeUndefined();
+            expect(t.get("t")).toBe(4);
+            expect(t.get("tt")).toBe(3);
+
+        });
+
+        it("support sub attribute name", function () {
 
             function a() {
                 a.superclass.constructor.apply(this, S.makeArray(arguments));
@@ -77,11 +147,11 @@ KISSY.use("base", function(S, Base) {
                 tt:{
                     // do not  use this in real world code
                     // forbid changing value in getter
-                    getter:function(v) {
+                    getter:function (v) {
                         v.x.y++;
                         return v;
                     },
-                    setter:function(v) {
+                    setter:function (v) {
                         v.x.y++;
                         return v;
                     }
@@ -98,12 +168,12 @@ KISSY.use("base", function(S, Base) {
                     }
                 }
             });
-            var ret = [],getterVal,setterVal;
-            t.on("beforeTtChange", function(e) {
+            var ret = [], getterVal, setterVal;
+            t.on("beforeTtChange", function (e) {
                 ret.push(e.prevVal.x.y);
                 ret.push(e.newVal.x.y);
             });
-            t.on("afterTtChange", function(e) {
+            t.on("afterTtChange", function (e) {
                 ret.push(e.prevVal.x.y);
                 ret.push(e.newVal.x.y);
             });
@@ -115,35 +185,35 @@ KISSY.use("base", function(S, Base) {
 
             expect(t.get("tt.x.y")).toBe(5);
 
-            expect(ret).toEqual([4,3,4,4]);
+            expect(ret).toEqual([4, 3, 4, 4]);
         });
 
-        it("should fire *Change once for set({})", function() {
+        it("should fire *Change once for set({})", function () {
             function a() {
                 a.superclass.constructor.apply(this, S.makeArray(arguments));
             }
 
             S.extend(a, Base);
 
-            var aa = new a({x:1,y:{z:1}}),
+            var aa = new a({x:1, y:{z:1}}),
                 ok = 0,
                 afterAttrChange = {};
 
-            aa.on("*Change", function(e) {
-                expect(e.newVal).toEqual([11,{z:22}]);
-                expect(e.prevVal).toEqual([1,{z:1}]);
-                expect(e.attrName).toEqual(["x","y"]);
-                expect(e.subAttrName).toEqual(["x","y.z"]);
-                ok ++;
+            aa.on("*Change", function (e) {
+                expect(e.newVal).toEqual([11, {z:22}]);
+                expect(e.prevVal).toEqual([1, {z:1}]);
+                expect(e.attrName).toEqual(["x", "y"]);
+                expect(e.subAttrName).toEqual(["x", "y.z"]);
+                ok++;
             });
-            aa.on("afterXChange", function(e) {
+            aa.on("afterXChange", function (e) {
                 expect(e.attrName).toBe("x");
                 expect(e.newVal).toBe(11);
                 expect(e.prevVal).toBe(1);
                 expect(e.subAttrName).toBe("x");
                 afterAttrChange.x = 1;
             });
-            aa.on("afterYChange", function(e) {
+            aa.on("afterYChange", function (e) {
                 expect(e.attrName).toBe("y");
                 expect(e.newVal).toEqual({z:22});
                 expect(e.prevVal).toEqual({z:1});