+++ src/ua/src/ua.js
@@ -1,7 +1,8 @@
 /**
  * @fileOverview ua
  */
-KISSY.add("ua", function(S,UA) {
+KISSY.add("ua", function (S, UA) {
+    S.UA = UA;
     return UA;
 }, {
     requires:["ua/extra"]